requirements.txt
pydantic~=2.10.6
fastapi~=0.115.7
httpx~=0.28.1
SQLAlchemy~=2.0.37
asyncpg~=0.30.0
greenlet~=3.1.1
alembic~=1.14.1
psycopg2~=2.9.10
APScheduler~=3.11.0
requests~=2.32.2
python-dotenv~=1.0.1
PyJWT~=2.10.1
python-multipart~=0.0.20
aiogram~=3.17.0
websockets~=14.2
uvicorn~=0.34.0

faiss-cpu
langchain==0.3.15
numpy==1.25.1
python-dotenv==1.0.1
sentence_transformers==3.3.1
torch==2.0.1
transformers==4.48.1
yandex_cloud_ml_sdk==0.2.0
